Adding --profile flag back, moving GLOBAL_ARGS to class; Closes #416
